refactor: separate deployment methods into reader/writer services
refs #1272
